Added OverviewPanel to make updating the overview panels in the model tab easier. Added GlobalSeqWizard for an easy way to make GlobalSeqs

+++ modelstudio/src/com/hiveworkshop/rms/editor/model/util/ModelFactory/IdObjectFactory.java
@@ -211,7 +211,7 @@ public class IdObjectFactory {
 
 		loadObject(ribbonEmitter, mdlxEmitter, model);
 
-		ribbonEmitter.setTextureSlot(mdlxEmitter.textureSlot);
+		ribbonEmitter.setTextureSlot((int) mdlxEmitter.textureSlot);
 		ribbonEmitter.setHeightAbove(mdlxEmitter.heightAbove);
 		ribbonEmitter.setHeightBelow(mdlxEmitter.heightBelow);
 		ribbonEmitter.setAlpha(mdlxEmitter.alpha);
